+++ src/indra_cogex/client/enrichment/signed.py
@@ -10,7 +10,7 @@ import pandas as pd
 import pystow
 import scipy.stats
 
-from indra_cogex.client.enrichment.gene_list import _collect_pathways
+from indra_cogex.client.enrichment.discrete import _collect_pathways
 from indra_cogex.client.neo4j_client import Neo4jClient
 
 HERE = Path(__file__).parent.resolve()
